Refactor message div shape based on topNoneRounded parameter.
This commit modifies the code to dynamically adjust the shape of the message div based on the topNoneRounded parameter.
The topNoneRounded parameter is a list with two boolean values.
The 0th index indicates whether there is a previous message (prev_message),
 and the 1st index indicates whether there is a next
  message (next_message).

The following changes have been made:

If next_message is True, the message div's shape is set to either rounded-ee-3xl or rounded-es-3xl based on the message's position (left or right).
If prev_message is True, the message div's shape is set to either rounded-se-3xl or rounded-ss-3xl based on the message's position (left or right).
These modifications ensure that the message div's shape is adjusted appropriately based on the presence of previous or next messages from the same user within a minute.